requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-crispy-forms==1.7.2
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.7.1
gunicorn==19.9.0
psycopg2==2.8.3
python-dotenv==0.10.3
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
